Create README.md
This is compulsory reading for LexDAO mentors in understanding your role (and responsibilities) towards the apprentices in the 4 wk program. The goal is to give a decent experience (as compared with law firm internship or judicial chamber clerk) so people get a sense of whether being a legal engineer is a career choice.